Renamed a file name to match its exported funtion name.
